Qualifying code

+++ robockey/localization.h
@@ -1,13 +1,13 @@
 #ifndef localization__
 #define localization__
 
-char get_location(unsigned int* arr, float* result);
+char get_location(unsigned int* arr, double* result);
 // FUNCTIONALITY:
 // Figure out position and orientation of robot
 //
 // TAKES:
 // arr: pointer to a 12-element unsigned int buffer
-// result: pointer to a 3-element float buffer
+// result: pointer to a 3-element double buffer
 //
 // RETURNS:
 // whatever m_wii_read returns
